Update MidiDataCalculations.h
Changed velocity from int to float because JUCE has a noteOn() that takes velocity as a float.

+++ Source/MidiDataCalculations.h
@@ -21,7 +21,7 @@ public:
     {
         float level = std::max(0.0f, std::min(levelIn, 1.0f));  // clamps from 0.0 to 1.0
         
-        velocity = (int)jmap(level, 0.0f, 1.0f, 0.0f, 127.0f);
+        velocity = jmap(level, 0.0f, 1.0f, 0.0f, 127.0f);
     }
     
     void setNoteOnOff(int noteNumber)
@@ -48,6 +48,6 @@ private:
     int previousNote = 0;
     
     // velocity variables
-    int velocity;
+    float velocity;
     
 };
